// src/api/model/friendDto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * PulseChat.API
 * PulseChat REST API
 * OpenAPI spec version: v1
 */
import type { FriendshipStatus } from './friendshipStatus';

export interface FriendDto {
  friendshipId?: string;
  userId?: string;
  /** @nullable */
  username?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  status?: FriendshipStatus;
  isRequester?: boolean;
  createdAt?: string;
}